chore(source): update build-logic/convention/src/main/kotlin/CMPFeatureConventionPlugin.kt cmp-android/build.gradle.kts cmp-android/src/main/kotlin/cmp/android/app/MainActivity.kt
