+++ build-logic/convention/src/main/kotlin/CMPFeatureConventionPlugin.kt
@@ -56,7 +56,7 @@ class CMPFeatureConventionPlugin : Plugin<Project> {
                 add("commonMainImplementation", project(":core-base:designsystem"))
                 add("commonMainImplementation", project(":core:data"))
                 add("commonMainImplementation", project(":core-base:designsystem"))
-                add("commonMainImplementation", project(":core:analytics"))
+                add("commonMainImplementation", project(":core:firebase"))
 
                 add("commonMainImplementation", libs.findLibrary("koin.compose").get())
                 add("commonMainImplementation", libs.findLibrary("koin.compose.viewmodel").get())
