+++ cmp-android/build.gradle.kts
@@ -105,7 +105,7 @@ dependencies {
 
     implementation(projects.cmpShared)
     implementation(projects.core.ui)
-    implementation(projects.core.analytics)
+    implementation(projects.core.firebase)
     implementation(projects.core.model)
     implementation(projects.core.data)
     implementation(projects.core.datastore)
