--- core-base/analytics/src/androidDemo/kotlin/kpt/core/base/analytics/di/AnalyticsModule.kt
@@ -1,21 +1,0 @@
-/*
- * Copyright 2025 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
- */
-package kpt.core.base.analytics.di
-
-import org.koin.core.module.Module
-import org.koin.core.module.dsl.singleOf
-import org.koin.dsl.bind
-import org.koin.dsl.module
-import kpt.core.base.analytics.AnalyticsHelper
-import kpt.core.base.analytics.StubAnalyticsHelper
-
-actual val analyticsModule: Module = module {
-    singleOf(::StubAnalyticsHelper) bind AnalyticsHelper::class
-}
