--- core-base/analytics/src/androidMain/AndroidManifest.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-    Copyright 2023 Mifos Initiative
-
-    This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
-    If a copy of the MPL was not distributed with this file,
-    You can obtain one at https://mozilla.org/MPL/2.0/.
-
-    See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
--->
-<manifest />
