--- core-base/analytics/src/firebaseMain/kotlin/kpt/core/base/analytics/FirebaseAnalyticsHelper.kt
@@ -1,37 +1,0 @@
-/*
- * Copyright 2025 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
- */
-package kpt.core.base.analytics
-
-import dev.gitlive.firebase.analytics.FirebaseAnalytics
-import dev.gitlive.firebase.analytics.logEvent
-
-class FirebaseAnalyticsHelper(
-    private val firebaseAnalytics: FirebaseAnalytics,
-) : AnalyticsHelper {
-
-    override fun logEvent(event: AnalyticsEvent) {
-        firebaseAnalytics.logEvent(event.type) {
-            for (extra in event.extras) {
-                param(
-                    key = extra.key.take(40),
-                    value = extra.value.take(100),
-                )
-            }
-        }
-    }
-
-    override fun setUserProperty(name: String, value: String) {
-        firebaseAnalytics.setUserProperty(name.take(24), value.take(36))
-    }
-
-    override fun setUserId(userId: String) {
-        firebaseAnalytics.setUserId(userId)
-    }
-}
