+++ core/firebase/.gitignore
