+++ core/firebase/README.md
