+++ core/firebase/src/androidMain/AndroidManifest.xml
