+++ gradle/libs.versions.toml
@@ -103,8 +103,8 @@ koinAnnotationsVersion = "4.2.2"
 
 
 # KMP Toolkit Libraries
-# 3.5.9: cmp-network-monitor accuracy overhaul + Android seed fix (MobileByteLabs/KmpToolkit#150).
-kmptoolkit = "3.5.9"
+# 3.5.11: cmp-firebase analytics+crashlytics engine + cmp-firebase-compose companion (KmpToolkit dev).
+kmptoolkit = "3.5.11"
 
 # CMP Libraries
 compose-plugin = "1.11.1"
@@ -370,6 +370,9 @@ cmp-intent-launcher = { module = "io.github.mobilebytelabs:cmp-intent-launcher",
 cmp-share = { module = "io.github.mobilebytelabs:cmp-share", version.ref = "kmptoolkit" }
 cmp-network-monitor = { module = "io.github.mobilebytelabs:cmp-network-monitor", version.ref = "kmptoolkit" }
 cmp-network-monitor-compose = { module = "io.github.mobilebytelabs:cmp-network-monitor-compose", version.ref = "kmptoolkit" }
+# cmp-firebase — unified Firebase (Analytics + Crashlytics) engine + Compose companion.
+cmp-firebase = { module = "io.github.mobilebytelabs:cmp-firebase", version.ref = "kmptoolkit" }
+cmp-firebase-compose = { module = "io.github.mobilebytelabs:cmp-firebase-compose", version.ref = "kmptoolkit" }
 
 firebase-analytics = { group = "com.google.firebase", name = "firebase-analytics" }
 firebase-bom = { group = "com.google.firebase", name = "firebase-bom", version.ref = "firebaseBom" }
